fix typo for verity on user builds
I might be misunderstanding the intention of the if/then statement, but these changes will result in system.img with a verity signature.  The non-verity fstab is moved to the else statement so that it is only defined once.

Users with an unlocked bootloader won't notice since they will get the orange warning.  I have a locked bootloader and these changes will result in the yellow warning with key signature.

If this change is correct, it can also be pushed to the Huawei angler and other devices that share this code.